requirements.txt
psycopg2-binary
python-dotenv
django-tailwind
Pillow